[feat] 술 등록, 수정, 삭제, 상세 조회 완료.

+++ server/src/main/java/com/bside/bside_311/dto/AddPostResponseDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Getter
 public class AddPostResponseDto {
   @Schema(example = "1", description = "게시글 번호")
-  private Integer postNo;
+  private Long postNo;
 }
